Create 238. Product of Array Except Self
Since division is not allowed, prefix and suffix product has to be taken out (elements before and after the element that needs to be avoided). Another way to do it is using two other arrays for the prefix and suffix product but this one eliminates that and saves space. It took me a minute to understand and I had to run it on paper to properly get it.